improve debuggability of block-less assertions
This allows block-less assertions to be debugged with the same level of
accuracy as normal block assertions.  We do this by keeping track of
Ruby's bindings and grabbing the one for the block-less assertion call.

    x = 1
    y = 3
    T { x == y }  # normal block assertion
    T x == y      # a block-less assertion

In both cases, you will be able to inspect the local variables x and y!

This commit wholly replaces 69b4c39abec31254eedbc08a95625cd379702155.