remove ruby-debug integration; use IRB only
ruby-debug is not very useful for interactive debugging in Dfect because
we only start the debugger at (and within) the point of failure.  As a
result, the debugger knows nothing about the call stack that led up to
the failure, so we cannot use it to visit previous stack frames.

This essentially reduces the capability of ruby-debug to that of a basic
interactive shell---which can only inspect the stack frame at the point
of failure.  Since IRB does a better job at being an interactive shell,
I am hereby removing ruby-debug integration from Dfect.  YAGNI/KISS FTW!