adjust no-binding line number to nearest below
Sometimes there is no binding for the line number given in the
backtrace, so we must adjust it to one that actually has a binding.

This problem occurs because line numbers reported in backtraces
sometimes do not agree with those observed by set_trace_func(),
particularly in method calls that span multiple lines:
set_trace_func() will consistently observe the ending parenthesis
(last line of the method call) whereas the backtrace will oddly
report a line somewhere in the middle of the method call.

NOTE: I chose to adjust the imprecise line to the nearest one BELOW
it.  This might not always be correct because the nearest line below
could belong to a completely different scope, like a new class.